Extracted rgb values from rgb integer

+++ pcl_demos/src/cylinder_segmentation.cpp
@@ -144,7 +144,7 @@ cloud_cb (const pcl::PCLPointCloud2ConstPtr& cloud_blob)
         uint8_t r = (rgb >> 16) & 0x0000ff;
         uint8_t g = (rgb >> 8)  & 0x0000ff;
         uint8_t b = (rgb)     & 0x0000ff;
-        std::cerr << r << "," << g << "," << b << std::endl;
+        std::cerr << (int)r << "," << (int)g << "," << (int)b << std::endl;
     }
